+++ src/main/java/com/finance/treasuretracker/service/AccountServiceImpl.java
@@ -23,7 +23,7 @@ public class AccountServiceImpl implements AccountServiceInterface {
     }
 
     @Override
-    public Account getAccountById(Long id) {
+    public Account getAccountById(Integer id) {
         return accountRepository.findById(id).orElse(null);
     }
 
@@ -38,7 +38,7 @@ public class AccountServiceImpl implements AccountServiceInterface {
     }
 
     @Override
-    public void deleteAccount(Long id) {
+    public void deleteAccount(Integer id) {
         accountRepository.deleteById(id);
     }
 }
